create bapsf_motion.utils and move ipv4_pattern and SimpleSignal to it
